Add tuple to example config
