+++ apps/consumer-client/src/podExampleConstants.ts
@@ -40,6 +40,12 @@ export const EXAMPLE_GPC_CONFIG = `{
         }
       }
     }
+  },
+  "tuples": {
+    "tuple0": {
+      "entries": ["pod0.A", "pod0.E"],
+      "isMemberOf": "admissiblePairs"
+    }
   }
 }`;
 
@@ -51,5 +57,12 @@ export const EXAMPLE_MEMBERSHIP_LISTS = `{
     9,
     "something",
     18711405342588116796533073928767088921854096266145046362753928030796553161041
+  ],
+  "admissiblePairs": [
+    [0,0],
+    [5,6],
+    [123,123],
+    ["zero", "zero"],
+    [0, "one"]
   ]
 }`;
