+++ packages/lib/gpc/src/gpcTypes.ts
@@ -1,11 +1,4 @@
-import {
-  POD,
-  PODEntries,
-  PODMembershipLists,
-  PODName,
-  PODValue,
-  PODValueTuple
-} from "@pcd/pod";
+import { POD, PODEntries, PODName, PODValue, PODValueTuple } from "@pcd/pod";
 import { Identity } from "@semaphore-protocol/identity";
 import { Groth16Proof } from "snarkjs";
 
@@ -15,6 +8,11 @@ import { Groth16Proof } from "snarkjs";
  * checked by {@link POD_NAME_REGEX}.
  */
 export type PODEntryIdentifier = `${PODName}.${PODName}`;
+
+/**
+ * Optional set of lists for checking POD entry (or tuple) value membership.
+ */
+export type PODMembershipLists = Record<PODName, PODValue[] | PODValueTuple[]>;
 
 // Single source of truth for tuple prefix (used internally).
 // This should not be a valid {@link PODName} to avoid name clashes.
@@ -426,7 +424,7 @@ export type GPCRevealedClaims = {
    * may be primitive (i.e. of type PODValue) or tuples (represented as
    * PODValueTuple = PODValue[]).  Each list must be non-empty.
    */
-  membershipLists?: Record<PODName, PODValue[] | PODValueTuple[]>;
+  membershipLists?: PODMembershipLists;
 
   /**
    * If this field is set, it matches the corresponding field in
