+++ packages/lib/pod/src/podTypes.ts
@@ -100,8 +100,3 @@ export type PODRawValueTuple = PODRawValue[];
  * ensure all entries are valid.
  */
 export type PODEntries = Record<PODName, PODValue>;
-
-/**
- * Optional set of lists for checking POD entry (or tuple) value membership.
- */
-export type PODMembershipLists = Record<PODName, PODValue[] | PODValueTuple[]>;
